server/python/requirements.txt
flask==3.0.2
flask-cors==4.0.0
torch==2.6.0
torchvision==0.21.0
Pillow==9.5.0
werkzeug==3.0.1
numpy>=1.24.0
python-dotenv==1.0.0
requests==2.31.0
uuid==1.30